refactor: directory rename: network to message; update server module's __all__
- Update import paths for Message and MessageType to the new message module
