+++ AloneChat/core/__init__.py
@@ -1,4 +1,4 @@
-from .network.protocol import Message, MessageType
+from .message.protocol import Message, MessageType
 from .plugin import Plugin, PluginManager
 
 __all__ = ['Message', 'MessageType', 'Plugin', 'PluginManager']
