+++ AloneChat/core/server/__init__.py
@@ -1,2 +1,5 @@
-# 预留服务端扩展接口
-pass
+from .manager import UserSession, SessionManager, WebSocketManager
+
+__all__ = [
+    'WebSocketManager'
+]
